model Email VerificationDAO
