+++ src/foam/nanos/auth/email/EmailVerificationDAO.js
@@ -1,0 +1,52 @@
+/**
+ * @license
+ * Copyright 2018 The FOAM Authors. All Rights Reserved.
+ * http://www.apache.org/licenses/LICENSE-2.0
+ */
+foam.CLASS({
+  package: 'foam.nanos.auth.email',
+  name: 'EmailVerificationDAO',
+  extends: 'foam.dao.ProxyDAO',
+
+  javaImports: [
+    'foam.core.X',
+    'foam.dao.DAO',
+    'foam.nanos.auth.User'
+  ],
+
+  axioms: [
+    {
+      name: 'javaExtras',
+      buildJavaClass: function(cls) {
+        cls.extras.push(
+          `
+            EmailTokenService emailToken;
+
+            public EmailVerificationDAO(X x, DAO delegate) {
+              setX(x);
+              setDelegate(delegate);
+              emailToken = (EmailTokenService) x.get("emailToken");
+            }
+          `
+        );
+      }
+    }
+  ],
+
+  methods: [
+    {
+      name: 'put_',
+      javaCode: `
+        boolean newUser = getDelegate().find(((User) obj).getId()) == null;
+
+        // send email verification if new user
+        User result = (User) super.put_(x, obj);
+        if ( result != null && newUser && ! result.getEmailVerified() ) {
+          emailToken.generateToken(x, result);
+        }
+    
+        return result;
+      `
+    }
+  ]
+});
